frg_compSVI

Function used to call the IDL functiosn used for computation of SNDVI yearly images (frg_compute_med_SVI.pro)

frg_compSVI(opts, force_update)

Arguments

opts
list of options passed from frg_fullprocessing()

force_update
logical If TRUE, recreate the ROI file even if it already exist, default: FALSE

Value

'DONE' if all went OK, otherwise error message to be treated by 'try.catch'

Details

This function is used to call the IDL function used for computation of SDVI and opts$SNDVI yearly images. It also calls the IDL routines needed to create the ENVI burnt areas ROI, and the ENVI burnt areas MASKS (both total and eroded!). Yearly opts$SRDVI (or opts$SNDVI) images are computed starting from the corresponding Yearly Average mosaics. Computation of the scaled opts$indexes for each pixel is done as follows:

The pixels belonging to the same class of the target included in a window of width specified by the user centred on the target are identified on the basis of the CLC00 recoded map;

Pixels showing NODATA values in the input VI image, or corresponding to areas classified as 'burned' on the input ROI files are removed from the above selection

Remaining pixels (i.e., same class, unburned and with good data) are used to compute the MOMENTS of the distribution of VI values of the target CLC class in the window centred on the target pixel

Computed MOMENTS are use to determine an ESTIMATE of the 5th, 50th (median) 95th percentile of the distribution, using the Cornish-Fisher fourth-order Expansion opts$method ("http://www.nematrian.com/R.aspx?p=CornishFisherDerivation)

Scaled opts$indexes for the target pixel is computed exploiting this opts$method:

'Percentage Deviation to Kernel Median': For each pixel, the algorithm identifies all pixels of the image corresponding to the same CLC class contained in a window of width opts$nker then computes the median of the distribution of values of the analyzed Vegetation opts$index in the selecteds pixels.

Scaled opts$index for the pixel are then computed as: sVI = 100 * (VI_pix - VI_median)/(VI_median).

(For further details, see the documentation for the IDL function frg_compute_med_SVI_batch)