frg_createmeta

Create ENVI META files allowing access to time series of Med_SVI files

frg_createmeta(opts, out_files, force_update = FALSE)

Arguments

opts
list of options passed from frg_fullprocessing()

out_files
character names of raster files to be used to create the META

force_update
logical If TRUE, recreate the ROI file even if it already exist, default: FALSE

Value

The function is called for its side effects