frg_main

Main Function of the Fire Regeneration Monitoring Tool

frg_main(force_update = FALSE)

Arguments

force_update
logical If TRUE, intermediate processing outputs (e.g., creation of burnt areas shapefile, computation of SVI, etc. are re-run even if intermecdiate processing outputs are already present in the specified output folders. Useful to make a "clean run" of the tool, Default: FALSE)

Value

Opens the Main GUI of the Tool and allow selection of the processing step to be run.

Details

This is the main function of the Fire Regeneration Monitoring Tool. It builds the main GUI and handles events corresponding to the different buttons, calling the corresponding functions. Also loads all required libraries, sources necessary routines and sets the main processing variables. The main processing functions of the FRG Tool can be accessed from the Main GUI. In particular:

Create Burned Area Shapefile: Allows creating the input burnt areas shapefile, starting from data present in the EFFIS database (esposito).

Analyse burned ares: Allows performing all the processing needed for the extraction and analysis of time series of rescaled indexes for the burnt areas present in the input shapefile.

Update Oracle tables: Update the EFFIS Oracle database tables regarding Regeneration Monitoring using the results of the latest processing.