frg_modproc

Function used to download, mosaic and preprocess MODIS data

frg_process_modis(opts, UI_check, max_UI, force_update)

Arguments

opts
string Folder where the original and preprocessed image will be stored

UI_check
flag If = 1 then a chack on pixel usefulness index is done while determining pixels to be used to compute average NDVI. default = 1

max_UI
Max UI value kept when averaging MODIS data. default = 5

force_update
DESCRIPTION , Default: FALSE

Details

Function used to download, mosaic and preprocess MODIS data.

Download and Mosaicing: Images covering Europe for end July - mid august of each selected year are downloaded, mosaiced and reprojected Resulting images are saved in TIFF format in the Originals subfolder of the output folder selected by the user. Separate subfolders are used to store images related to NDVI, RED and NIR reflectance, QA (Quality Assurance) and Pixel Reliability.

Preprocessing: The preprocessing steps conducted are the following: a. MODIS UI (Usefulness Index) values are extracted from the MODIS QA (Quality Assurance) files c. The average yearly values for DVI and NDVI are computed starting from the images of DOY 209 and 225. NODATA and bad quality data#' are automatically removed