FRG_Comp_Sig_Matrix

Function used to perform the Wilcoxon test of significant differences, using differnt number of 'before fire years' as reference. The function creates and saves matrixes of p-values obtained using differnt number of before fire years. Matrix is saved within the 'p-matrixes' subfolder of the folder containing the final output file. Additional info necessary to plot the time series (i.e., boxplots, objectids of fires, fireyears, ecc) are also saved in the same file (plot_stat data frame).

frg_wilcox(in_file, out_file_pmat, opts)

Arguments

in_file
string File of scaled VI Time Series extracted for single-fire areas

out_file_pmat
character path to the file to be used to store the p-values matrix

opts
list of options passed from frg_fullprocessing()

Value

Saves a p_matrix.RData file containing the p-values matrix and the 'plot_stats' data frame in the 'p_matrixes' subfolder